refactor: delete dead frontend components and their source-text test pins
Remove unreferenced sources that only survived because tests read their
text: atlas/ComputeJobs.tsx, atlas/OpenScienceFileTree.tsx,
components/dialog-select-model.{tsx,css}, components/model-tooltip.tsx,
and ui/components/session-review.{tsx,css} (plus its index.css import).

Trim the tests that pinned those files. ComputeJobs.test.ts becomes
ComputeJobsAPI.test.ts and keeps the createComputeJobsAPI, serial,
stableJobs, and ComputeJobsAPI.ts path assertions.
