+++ frontend/workspace/src/pages/project-navigation.test.ts
@@ -103,16 +103,9 @@ describe("project-ID navigation wiring", () => {
   })
 
   test("keeps linked-worktree sessions and file surfaces on the selected directory", async () => {
-    const [explorer, tree, preview] = await Promise.all([
-      read("../atlas/FileExplorer.tsx"),
-      read("../atlas/OpenScienceFileTree.tsx"),
-      read("../atlas/FilePreview.tsx"),
-    ])
+    const [explorer, preview] = await Promise.all([read("../atlas/FileExplorer.tsx"), read("../atlas/FilePreview.tsx")])
 
     expect(explorer).toContain("props.file.directory || sdk.directory")
-    expect(tree).toContain(
-      'const directory = () => sdk.directory || sync.data.path.directory || sync.project?.worktree || ""',
-    )
     expect(preview).toContain("props.directory || production?.directory || sync?.data.path.directory")
   })
 })
